Created tests for money transfer

+++ src/main/java/org/example/exception/InsufficientFundsException.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public class InsufficientFundsException extends ClientSideException {
 
     public InsufficientFundsException(Long userId, BigDecimal amount) {
-        super(String.format("User '%s' has insufficient amount of money to transfer '%d'", userId, amount));
+        super(String.format("User '%s' has insufficient amount of money to transfer '%d'", userId, amount.intValue()));
     }
 }
